[Be/feat/clover] 친밀도/모래알 증가 메서드, 클로버 엔티티

+++ dolharubang/src/main/java/com/dolharubang/domain/dto/request/MemberReqDto.java
@@ -9,8 +9,8 @@ public class MemberReqDto {
     private String memberEmail;
     private String nickname;
     private String birthday;
-    private Long sands;
-    private Long totalLoginDays;
+    private int sands;
+    private int totalLoginDays;
     private String imageBase64;
     private String spaceName;
